create connections between nodes
To create connections between Node objects, it was elected that each Node store it's own connecting partners. This information of which nodes are connected to which nodes is necessary to allow for singular creation of a connection object between the two nodes (i.e. "create a connection if the nodes are within a certain distance and there does not already exist a connection between the two"). This method was elected for instead of some form of map structure which would use the two objects as a key and then store the connection object between them as the value. While this implementation might be faster it seems overly convoluted and would require a hashing algorithm to use the combination of the objects as key identifiers. Instead each Node will contain a list of connections, when that node is being check against another node for an existing connection a linear search will be done to determine if that node is indeed a connecting node.