refactor node connections and connection drawing
Previously connections were made and maintained dynamically. They were created once particles were close enough and each particle tracked its partners. Additionally they were removed the the connections_ vector when particles were too far away (this proved to be buggy however as removal from the connections_ vector did not remove the connection in the connecting partners vector located in the individual particles). Instead of this format, it was opted for a simpler approach. Instead of dynamically generating connections, a connection obj is made between every pair of nodes. The only thing that changes is when the nodes are close enough, those connections get displayed on the screen. This switch is a trade-off. I am exchanging possible efficiency increases for simplicity. In a hypothetical scenario of 100 particles we could assume that each particle (using the previous method) would maintain on average 5 connections. This would mean we would have to track and update ~500 connections in total. With the new implementation 100 particles will have 4950 total connections (with most not being visible at one time). The simplicity gains at this stage in the project are much more valuable than the efficiency gains. If there is ever a point where processing that many connections become laborious, the switch back can always be made.